requirements.txt
flask==2.3.3
python-dotenv==1.0.0
twelvelabs
qdrant-client==1.7.0
Werkzeug==2.3.7
requests==2.31.0
pillow==10.0.0
Werkzeug